refactor(FlyingWardrobe): extract patrol movements in a new script 'PatrolMover'
