+++ Assets/Scripts/Ennemies/FlyingWardrobe/FlyingWardrobeLogic.cs
@@ -2,16 +2,11 @@ using Assets.Scripts.Gameplay;
 using Assets.Scripts.Utils;
 using UnityEngine;
 
-[RequireComponent(typeof(FlyingWardrobeAnimations)), RequireComponent(typeof(PeriodicAttack))]
+[RequireComponent(typeof(FlyingWardrobeAnimations)), RequireComponent(typeof(PeriodicAttack)), RequireComponent(typeof(PatrolMover))]
 public class FlyingWardrobeLogic : MonoBehaviour, IAttackCondition
 {
     [Header("Components References")]
     public FlyingWardrobeAnimations animations;
-
-    [Header("Movements Parameters")]
-    public float travellingDistance;
-    public float velocity;
-    public FacingDirection initialFacingDirection;
 
     [Header("Attack Parameters")]
     public int plateCount = 3;
@@ -25,38 +20,24 @@ public class FlyingWardrobeLogic : MonoBehaviour, IAttackCondition
     public float plateDespawnDistance;
 
     [Header("States")]
-    private bool canMove;
     private Vector3 currentPosition;
-    private Vector3 initialPosition;
-    private FacingDirection facingDirection;
     private Vector3 playerPos;
-    private bool moveAbovePlayer;
 
     private void FindComponents()
     {
         this.animations = this.animations.GetComponent<FlyingWardrobeAnimations>();
     }
 
-    private void InitializeStates()
-    {
-        this.initialPosition = this.transform.position;
-        this.currentPosition = this.initialPosition;
-        this.canMove = true;
-        this.facingDirection = this.initialFacingDirection;
-        this.moveAbovePlayer = false;
-    }
-
     void Start()
     {
         this.FindComponents();
-        this.InitializeStates();
     }
 
     private Vector3 GeneratePlatePosition()
     {
         float xOffset = Mathf.Abs(this.plateOffsetRange.x);
         float yOffset = Mathf.Abs(this.plateOffsetRange.y);
-        var range = new Vector3(UnityEngine.Random.Range(-xOffset, xOffset), UnityEngine.Random.Range(-yOffset, yOffset), 0f);
+        Vector3 range = new(Random.Range(-xOffset, xOffset), Random.Range(-yOffset, yOffset), 0f);
         return this.transform.position + range;
     }
 
@@ -75,101 +56,6 @@ public class FlyingWardrobeLogic : MonoBehaviour, IAttackCondition
         return Mathf.Abs(this.currentPosition.x - this.playerPos.x) <= Mathf.Abs(this.spawnTriggerDistThresh);
     }
 
-    private void Move()
-    {
-        if (!this.canMove)
-        {
-            return;
-        }
-
-        RightOrLeft();
-
-        if (this.moveAbovePlayer)
-        {
-            this.currentPosition.x = Mathf.MoveTowards(this.currentPosition.x, this.playerPos.x, this.velocity * Time.deltaTime);
-        }
-        else if (this.facingDirection == FacingDirection.Right)
-        {
-            this.currentPosition.x += this.velocity * Time.deltaTime;
-        }
-        else
-        {
-            this.currentPosition.x -= this.velocity * Time.deltaTime;
-        }
-
-        var movement = new Vector3(this.currentPosition.x, this.transform.position.y, this.transform.position.z);
-        if (this.animations.IsInState("Flying"))
-        {
-            movement += this.animations.GetVerticalOffset();
-        }
-
-        this.transform.position = movement;
-    }
-
-    private void RightOrLeft()
-    {
-        if (this.moveAbovePlayer)
-        {
-            // Update facing direction based on the player position relative to enemy
-            if (this.currentPosition.x < this.playerPos.x - 0.05f)
-            {
-                this.facingDirection = FacingDirection.Right;
-            }
-            else if (currentPosition.x > playerPos.x + 0.05f)
-            {
-                this.facingDirection = FacingDirection.Left;
-            }
-        }
-        else
-        {
-            if (this.initialFacingDirection == FacingDirection.Right)
-            {
-                float targetX = this.initialPosition.x + this.travellingDistance;
-                if (this.currentPosition.x < this.initialPosition.x)
-                {
-                    this.facingDirection = FacingDirection.Right;
-                }
-                else if (currentPosition.x > targetX)
-                {
-                    this.facingDirection = FacingDirection.Left;
-                }
-            }
-            else
-            {
-                float targetX = this.initialPosition.x - this.travellingDistance;
-                if (this.currentPosition.x > this.initialPosition.x)
-                {
-                    this.facingDirection = FacingDirection.Left;
-                }
-                else if (this.currentPosition.x < targetX)
-                {
-                    this.facingDirection = FacingDirection.Right;
-                }
-            }
-        }
-    }
-
-    private void FaceDirection()
-    {
-        Vector3 currentScale = this.transform.localScale;
-        if (this.facingDirection == FacingDirection.Right && currentScale.x > 0)
-        {
-            currentScale.x = -Mathf.Abs(currentScale.x);
-            this.transform.localScale = currentScale;
-        }
-        else if (this.facingDirection == FacingDirection.Left && currentScale.x < 0)
-        {
-            currentScale.x = Mathf.Abs(currentScale.x);
-            this.transform.localScale = currentScale;
-        }
-    }
-
-    void Update()
-    {
-        Move();
-        FaceDirection();
-    }
-
     public void Attack()
     {
         this.SpawnPlates();
@@ -180,25 +66,12 @@ public class FlyingWardrobeLogic : MonoBehaviour, IAttackCondition
         return this.IsNearPlayer();
     }
 
-    private void OnTriggerStay2D(Collider2D collision)
+    public void OnMovementComputed(Vector3 movement)
     {
-        if (collision.CompareTag("Player"))
+        if (this.animations.IsInState("Flying"))
         {
-            this.moveAbovePlayer = true;
-            this.playerPos.x = collision.transform.position.x;
+            movement += this.animations.GetVerticalOffset();
         }
-        else
-        {
-            this.moveAbovePlayer = false;
-        }
-    }
-
-    private void OnTriggerExit2D(Collider2D collision)
-    {
-        if (collision.CompareTag("Player"))
-        {
-            this.canMove = true;
-            this.moveAbovePlayer = false;
-        }
+        this.transform.position = movement;
     }
 }
